+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [17.22.1]
 - [RefreshView] Removed work around for RefreshView for .NET 7 which causes a glitch for iOS CollectionView in .NET 8.
+- [ContextMenuItem] Made sure Item is a bindable property.
 
 ## [17.22.0]
 - Added 'FormattedText' property on TitleOptions